+++ src/pages/main/components/Carousel.tsx
